[Implementation] App is working. Need to test edge cases and improve error handling

+++ app/src/main/java/com/example/weatherappapolinsky/data/Clouds.kt
@@ -7,5 +7,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class Clouds(
     @SerialName("all")
-    var all: Int? = null
+    var all: Double? = null
 )
